Crate pwasm_utils

Modules

rules

stack_height
The pass that tries to make stack overflows deterministic, by introducing an upper bound of the stack size.

Enums

OptimizerError

PackingError
Pack error.

Statics

CALL_SYMBOL

CREATE_SYMBOL

RET_SYMBOL

Functions

externalize

externalize_mem

init_log
Intialize log with default settings

inject_gas_counter
Injects gas counter.

inject_runtime_type

optimize

pack_instance
If module has an exported "_create" function we want to pack it into "constructor". raw_module is the actual contract code ctor_module is the constructor which should return raw_module

shrink_unknown_stack

underscore_funcs

ununderscore_funcs